Introduce S3Action type for S3 actions
- Add ActionType and improve validation.

- Remove useless {policy,statement}.isValid

- Rename BucketPolicy`.isValid` to `.validate`

- Fix one-off typed constant AbortMultipartUploadAction to be untyped.

Fix a linter error in a a test (workers/workers_test.go)
